backend/requirements.txt
streamlit==1.36.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9

sentence-transformers==2.7.0
faiss-cpu==1.8.0

pypdf==4.2.0
pdfminer.six==20221105
python-docx==1.1.2
fpdf2==2.7.9
reportlab==4.4.3

python-dotenv==1.0.1
requests==2.32.4
httpx==0.27.2
